+++ pkgs/test_core/CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.4.14-dev
+## 0.4.14
 
 * Update the github reporter to output the platform in the test names when
   multiple platforms are used.